requirements.txt
fastapi
pandas
uvicorn
faker
google-cloud-storage
fsspec
gcsfs
smart_open
sqlalchemy
cloud-sql-python-connector
google-auth 
google-auth-oauthlib
google-auth-httplib2
apache-airflow
apache-airflow-providers-postgres
apache-airflow-providers-apache-spark
